Restructure and update CLI entry point to run_pageindex.py
